+++ src/components/grid_popover/grid_popover.ts
@@ -1,8 +1,8 @@
 import { Component } from "@odoo/owl";
 import { cellPopoverRegistry } from "../../registries";
-import { CellPopover } from "../../store/cell_popover";
-import { Store } from "../../store/dependency_container";
-import { useStore } from "../../store/store_hooks";
+import { CellPopover } from "../../stores/cell_popover";
+import { Store } from "../../store_engine/store";
+import { useStore } from "../../store_engine/store_hooks";
 import { Rect, SpreadsheetChildEnv } from "../../types";
 import { ClosedCellPopover, PositionedCellPopover } from "../../types/cell_popovers";
 import { ErrorToolTipPopoverBuilder } from "../error_tooltip/error_tooltip";
