+++ src/components/link/link_display/link_display.ts
@@ -2,9 +2,9 @@ import { Component } from "@odoo/owl";
 import { LINK_COLOR } from "../../../constants";
 import { toXC } from "../../../helpers";
 import { openLink, urlRepresentation } from "../../../helpers/links";
-import { CellPopover } from "../../../store/cell_popover";
-import { Store } from "../../../store/dependency_container";
-import { useStore } from "../../../store/store_hooks";
+import { CellPopover } from "../../../stores/cell_popover";
+import { Store } from "../../../store_engine/store";
+import { useStore } from "../../../store_engine/store_hooks";
 import { EvaluatedCell, Link, Position, SpreadsheetChildEnv } from "../../../types";
 import { CellPopoverComponent, PopoverBuilders } from "../../../types/cell_popovers";
 import { css } from "../../helpers/css";
