+++ src/store_engine/design.md
